separate localwrite module
